changing files to correct place
